Async s_rate / BCLK selection in i2s master
Could be done nicer, but for now does the job of giving as 192k sample
rate @32bits stereo I2S from 12.288 SAICLK
It will switch asynchronously to the SAICLK driven logic in the i2s
master and can cause serious glitches for at the point of switching
sample rates. But as this happens only at boot time or when changed by
user, a glitch should not be a problem.